MSVS 2015u3 warning

+++ modules/xfeatures2d/src/pct_signatures/grayscale_bitmap.cpp
@@ -110,7 +110,7 @@ namespace cv
                     }
                 }
                 // Prepare the preallocated contrast matrix for contrast-entropy computations
-                mCoOccurrenceMatrix.resize(1 << (mBitsPerPixel * 2));   // mCoOccurrenceMatrix size = maxPixelValue^2
+                mCoOccurrenceMatrix.resize((int)(1 << (mBitsPerPixel * 2)));   // mCoOccurrenceMatrix size = maxPixelValue^2
             }
